Restructuring miner dialogue adjustments
The miners used different state numbers for their mines-are-finished top-level states across platforms (and in the case
of BGEE, had none). Calling now just creates its own state in these circumstances and bypasses their normal dialogue
entirely while the sequence of helping Melicamp is active--given that it's a short and localized sequence, I don't
terribly mind hijacking the states here even though it's not the most mod-friendly thing to do.