Three fixes:
Accounting for black pearls in the Balduran's Seatower mod
Chatted with jastey for a bit about the black pearls in The Calling and in BST. In BST, there's one and it's a one-off quest
item, as opposed to the uncommon-but-not-unique way that TC treats them. As such, we're keeping them as separate items. BST will
re-label theirs as "Enshelled Black Pearl" and we're both changing to a unified description (BST adds one line to it to indicate
its uniqueness). I've added a reply for Thalantyr in which he indicates that the BST pearl is beautiful, but unusable, so that
we don't bork BST's quest. jastey will be adding a response to TC's black pearls to her quest-giver as well.

Keldath's unidentified pearls
If you ask Keldath for a black pearl and you have 'gems require ID" from Tweaks, Keldath gives you an unidentified pearl. As this
makes no sense, he'll now ID it first on EE games. I could do a workaround and stash it in inventory for non-EEs but frankly
it's just not worth the effort for an edge case, and opens it up to pickpocket shenanigans.

Updated logo in Github readme to the one usable on light or dark themes